Fix Several issues
1. Fix warning; Change All char* = "string" statement into char[] = "string" statement
2. In Print function, If the ast_XXX_string function called, then "free".

+++ src/verilog_writer.h
@@ -14,8 +14,8 @@ NAMESPACE_VERILOG_BEGIN
 // Verilog Writer Function
 void PrintVerilog( FILE* fout, verilog_source_tree * tree );
  
-int PrintExpression( FILE* fout, ast_expression* expression );
-int PrintPrimary( FILE* fout, ast_primary* primary );
+size_t PrintExpression( FILE* fout, ast_expression* expression );
+size_t PrintPrimary( FILE* fout, ast_primary* primary );
 
 NAMESPACE_VERILOG_END
